backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-genai
docling
